Finalizado cadastro de usuário e edição de perfil

+++ src/controllers/EmpresaController.js
@@ -2,8 +2,23 @@ const { connect } = require('../services/db')
 const { PrismaClient } = require('@prisma/client')
 const EmpresaService = require('../repositories/EmpresaService')
 
+function verificarAcesso(user) {
+    // 1 - AMD ROOT
+    // 2 - AMD INEA
+    // 3 - AMD ADM EMPRESAS
+    // 4 - FISCAIS
+    // 5 - FUNCIONARIOS
+    const listaUsuariosAutorizados = [1, 2, 3, 5]
+    return listaUsuariosAutorizados.some(item => item == user.user_tipo_id)
+}
+
 class EmpresaController {
     async criar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await EmpresaService.create(req.body)
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -13,6 +28,11 @@ class EmpresaController {
     }
 
     async listar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await EmpresaService.getAll()
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -22,6 +42,11 @@ class EmpresaController {
     }
 
     async exibir(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await EmpresaService.getById(Number(req?.params?.id))
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -32,6 +57,11 @@ class EmpresaController {
 
 
     async editar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const id = Number(req?.params?.id)
         const payload = req.body
         const dados = await EmpresaService.update(id, payload)
@@ -43,6 +73,11 @@ class EmpresaController {
     }
 
     async deletar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const id = Number(req?.params?.id)
         const dados = await EmpresaService.delete(id)
         if (!dados?.erro) {
