+++ src/controllers/PropriedadeController.js
@@ -1,9 +1,22 @@
 const { connect } = require('../services/db')
 const { PrismaClient } = require('@prisma/client')
 const PropriedadeService = require('../repositories/PropriedadeService')
-
+function verificarAcesso(user) {
+    // 1 - AMD ROOT
+    // 2 - AMD INEA
+    // 3 - AMD ADM EMPRESAS
+    // 4 - FISCAIS
+    // 5 - FUNCIONARIOS
+    const listaUsuariosAutorizados = [1, 3, 5]
+    return listaUsuariosAutorizados.some(item => item == user.user_tipo_id)
+}
 class UsuarioController {
     async criar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await PropriedadeService.create(req.body)
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -13,6 +26,11 @@ class UsuarioController {
     }
 
     async listar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await PropriedadeService.getAll()
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -22,6 +40,11 @@ class UsuarioController {
     }
 
     async exibir(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await PropriedadeService.getById(Number(req?.params?.id))
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -32,6 +55,11 @@ class UsuarioController {
 
 
     async editar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const id = Number(req?.params?.id)
         const payload = req.body
         const dados = await PropriedadeService.update(id, payload)
@@ -43,6 +71,11 @@ class UsuarioController {
     }
 
     async deletar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const id = Number(req?.params?.id)
         const dados = await PropriedadeService.delete(id)
         if (!dados?.erro) {
