+++ src/controllers/RepresentanteController.js
@@ -1,10 +1,24 @@
 const { connect } = require('../services/db')
 var prisma = require('../services/prisma')
 var RepresentanteService = require("../repositories/RepresentanteService")
+function verificarAcesso(user) {
+    // 1 - AMD ROOT
+    // 2 - AMD INEA
+    // 3 - AMD ADM EMPRESAS
+    // 4 - FISCAIS
+    // 5 - FUNCIONARIOS
+    const listaUsuariosAutorizados = [1, 3, 5]
+    return listaUsuariosAutorizados.some(item => item == user.user_tipo_id)
 
+}
 
 class ProprietarioController {
     async criar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await RepresentanteService.create(req.body)
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -14,6 +28,11 @@ class ProprietarioController {
     }
 
     async listar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const skip = Number(req?.query?.skip) || 0
         const take = Number(req?.query?.take) || 100
         const busca = req?.query?.busca || ""
@@ -27,6 +46,11 @@ class ProprietarioController {
     }
 
     async exibir(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const dados = await RepresentanteService.getById(Number(req?.params?.id))
         if (!dados?.erro) {
             res.status(200).send(dados)
@@ -36,6 +60,11 @@ class ProprietarioController {
     }
 
     async editar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const id = Number(req?.params?.id)
         const payload = req.body
         const dados = await RepresentanteService.update(id, payload)
@@ -47,6 +76,11 @@ class ProprietarioController {
     }
 
     async deletar(req, res) {
+        const user = req.user
+        if (!verificarAcesso(user)) {
+            res.status(401).send({ erro: true, msg: 'Acesso não autorizado' })
+            return
+        }
         const id = Number(req?.params?.id)
         const dados = await RepresentanteService.delete(id)
         if (!dados?.erro) {
